# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/operative": {
            "get": {
                "description": "list operative entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List operative entities",
                "operationId": "list-operative",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.Operative"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create operative",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create operative",
                "operationId": "create-operative",
                "parameters": [
                    {
                        "description": "Operative entity",
                        "name": "operative",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Operative"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Operative"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/operative/{id}": {
            "get": {
                "description": "get operative by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a operative entity by ID",
                "operationId": "get-operative",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Operative ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Operative"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update operative by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a operative entity by ID",
                "operationId": "update-operative",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Operative ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Operative entity",
                        "name": "operative",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.Operative"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.Operative"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get operative by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a operative entity by ID",
                "operationId": "delete-operative",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Operative ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ent.Examinationroom": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ExaminationroomQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ExaminationroomEdges"
                },
                "examinationroomName": {
                    "description": "ExaminationroomName holds the value of the \"examinationroomName\" field.",
                    "type": "string"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.ExaminationroomEdges": {
            "type": "object",
            "properties": {
                "fromexaminationroom": {
                    "description": "Fromexaminationroom holds the value of the fromexaminationroom edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Operativerecord"
                    }
                }
            }
        },
        "ent.Nurse": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the NurseQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.NurseEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "nurse_email": {
                    "description": "NurseEmail holds the value of the \"nurse_email\" field.",
                    "type": "string"
                },
                "nurse_name": {
                    "description": "NurseName holds the value of the \"nurse_name\" field.",
                    "type": "string"
                },
                "nurse_password": {
                    "description": "NursePassword holds the value of the \"nurse_password\" field.",
                    "type": "string"
                },
                "nurse_tel": {
                    "description": "NurseTel holds the value of the \"nurse_tel\" field.",
                    "type": "string"
                }
            }
        },
        "ent.NurseEdges": {
            "type": "object",
            "properties": {
                "nurseop": {
                    "description": "Nurseop holds the value of the nurseop edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Operativerecord"
                    }
                }
            }
        },
        "ent.Operative": {
            "type": "object",
            "properties": {
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the OperativeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.OperativeEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "operativeName": {
                    "description": "OperativeName holds the value of the \"operativeName\" field.",
                    "type": "string"
                },
                "operativeType": {
                    "description": "OperativeType holds the value of the \"operativeType\" field.",
                    "type": "string"
                }
            }
        },
        "ent.OperativeEdges": {
            "type": "object",
            "properties": {
                "fromoperative": {
                    "description": "Fromoperative holds the value of the fromoperative edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Operativerecord"
                    }
                }
            }
        },
        "ent.Operativerecord": {
            "type": "object",
            "properties": {
                "OperativeTime": {
                    "description": "OperativeTime holds the value of the \"OperativeTime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the OperativerecordQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.OperativerecordEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.OperativerecordEdges": {
            "type": "object",
            "properties": {
                "examinationroom": {
                    "description": "Examinationroom holds the value of the examinationroom edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Examinationroom"
                },
                "nurse": {
                    "description": "Nurse holds the value of the nurse edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Nurse"
                },
                "operative": {
                    "description": "Operative holds the value of the operative edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Operative"
                },
                "tool": {
                    "description": "Tool holds the value of the tool edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Tool"
                }
            }
        },
        "ent.Tool": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "Amount holds the value of the \"amount\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the ToolQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.ToolEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                },
                "toolName": {
                    "description": "ToolName holds the value of the \"toolName\" field.",
                    "type": "string"
                }
            }
        },
        "ent.ToolEdges": {
            "type": "object",
            "properties": {
                "fromtool": {
                    "description": "Fromtool holds the value of the fromtool edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Operativerecord"
                    }
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "authorizationUrl": "",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "authorizationUrl": "",
...
}